Code for initial release
This rather sizeable commit is the result of two rounds of prototyping
and an initial pull request into a private monorepo before we decided to
open source this as an independent repository and npm package.

I intend to open a pull request for this to invite a final round of
feedback before publishing the initial release and to ensure Github
Actions tests the build properly.